First working example (SHM)

+++ src/main/java/org/jgroups/shm/SharedMemoryBuffer.java
@@ -4,7 +4,6 @@ package org.jgroups.shm;
 import org.agrona.MutableDirectBuffer;
 import org.agrona.concurrent.*;
 import org.agrona.concurrent.ringbuffer.ManyToOneRingBuffer;
-import org.jgroups.logging.Log;
 import org.jgroups.util.DefaultThreadFactory;
 import org.jgroups.util.Runner;
 import org.jgroups.util.ThreadFactory;
@@ -15,47 +14,78 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
+import java.nio.file.OpenOption;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.Objects;
+import java.util.concurrent.TimeUnit;
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
 /**
- * Wraps {@link org.agrona.concurrent.ringbuffer.ManyToOneRingBuffer}.
+ * Wraps {@link org.agrona.concurrent.ringbuffer.ManyToOneRingBuffer}. Can be used for writing; reading is enabled
+ * by setting a consumer ({@link #setConsumer(Consumer)}).
  * @author Bela Ban
  * @since  1.0.0
  */
 public class SharedMemoryBuffer implements MessageHandler, Closeable {
-    protected final String        file_name;   // name of the shared memory-mapped file (e.g. /tmp/shm/uuid-1
-    protected final Consumer<ByteBuffer> consumer; // a received message calls consumer.receive();
-    protected FileChannel         channel;     // the memory-mapped file
-    protected UnsafeBuffer        buffer;
-    protected ManyToOneRingBuffer rb;
-    protected final Runner        runner;
-    protected IdleStrategy        idle_strategy=new BackoffIdleStrategy();
-    protected final Log           log;
+    protected final String         file_name;   // name of the shared memory-mapped file (e.g. /tmp/shm/uuid-1
+    protected BiConsumer<ByteBuffer,Integer> consumer; // a received message calls consumer.receive();
+    protected FileChannel          channel;     // the memory-mapped file
+    protected UnsafeBuffer         buffer;
+    protected ManyToOneRingBuffer  rb;
+    protected final Runner         runner;
+    protected IdleStrategy         idle_strategy;
+    protected boolean              delete_file_on_exit;
 
-    public SharedMemoryBuffer(String file_name, Consumer<ByteBuffer> c, int buffer_length, Log log) throws IOException {
+
+    public SharedMemoryBuffer(String file_name, int buffer_length, boolean create) throws IOException {
         this.file_name=file_name;
-        this.consumer=Objects.requireNonNull(c);
-        this.log=log;
-        init(buffer_length);
+        // idle stragegy spins, the yields, then parks between 1000ns and 64ms by default
+        idle_strategy=new BackoffIdleStrategy(BackoffIdleStrategy.DEFAULT_MAX_SPINS,
+                                              BackoffIdleStrategy.DEFAULT_MAX_YIELDS,
+                                              BackoffIdleStrategy.DEFAULT_MIN_PARK_PERIOD_NS,
+                                              BackoffIdleStrategy.DEFAULT_MAX_PARK_PERIOD_NS<<6);
+        init(buffer_length, create);
         ThreadFactory tf=new DefaultThreadFactory("runner", true, true);
         runner=new Runner(tf, String.format("shm-%s", file_name), this::doWork, null);
-        runner.start();
     }
+
 
     public SharedMemoryBuffer idleStrategy(IdleStrategy s) {idle_strategy=Objects.requireNonNull(s); return this;}
 
-    public boolean add(byte[] buf, int offset, int length) {
+    public SharedMemoryBuffer maxSleep(long m) {
+        long max_sleep_ns=TimeUnit.NANOSECONDS.convert(m, TimeUnit.MILLISECONDS);
+        idle_strategy=new BackoffIdleStrategy(BackoffIdleStrategy.DEFAULT_MAX_SPINS, BackoffIdleStrategy.DEFAULT_MAX_YIELDS,
+                                              BackoffIdleStrategy.DEFAULT_MIN_PARK_PERIOD_NS, max_sleep_ns);
+        return this;
+    }
+
+    public SharedMemoryBuffer deleteFileOnExit(boolean f)  {
+        if((delete_file_on_exit=f) == true) {
+            File tmp=new File(file_name);
+            tmp.deleteOnExit();
+        }
+        return this;}
+
+    public SharedMemoryBuffer setConsumer(BiConsumer<ByteBuffer,Integer> c) {
+        consumer=Objects.requireNonNull(c);
+        runner.start();
+        return this;
+    }
+
+    public void write(byte[] buf, int offset, int length) {
         int claim_index=rb.tryClaim(1, length);
         if(claim_index > 0) {
-            AtomicBuffer bb=rb.buffer();
-            bb.putBytes(claim_index, buf, offset, buf.length);
-            rb.commit(claim_index);
-            return true;
+            try {
+                AtomicBuffer bb=rb.buffer();
+                bb.putBytes(claim_index, buf, offset, length);
+                rb.commit(claim_index);
+            }
+            catch(Throwable t) {
+                rb.abort(claim_index);
+            }
         }
-        return false;
     }
 
     /**
@@ -69,9 +99,10 @@ public class SharedMemoryBuffer implements MessageHandler, Closeable {
 
     @Override
     public void onMessage(int msg_type, MutableDirectBuffer buf, int offset, int length) {
-        ByteBuffer bb=buf.byteBuffer();
-        bb.position(offset);
-        consumer.accept(bb);
+        if(msg_type != 1)
+            return;
+        ByteBuffer bb=buf.byteBuffer().position(offset);
+        consumer.accept(bb,length);
     }
 
     public void close() {
@@ -80,13 +111,17 @@ public class SharedMemoryBuffer implements MessageHandler, Closeable {
         tmp.delete();
     }
 
-    protected void init(int buffer_length) throws IOException {
+    protected void init(int buffer_length, boolean create) throws IOException {
         try {
-            File tmp=new File(file_name);
-            tmp.deleteOnExit();
-            channel=FileChannel.open(Paths.get(file_name),
-                                     StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE);
-                                     // StandardOpenOption.DELETE_ON_CLOSE);
+            if(delete_file_on_exit) {
+                File tmp=new File(file_name);
+                tmp.deleteOnExit();
+            }
+            OpenOption[] options=create?
+              new OpenOption[]{StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE} :
+              new OpenOption[]{StandardOpenOption.READ, StandardOpenOption.WRITE};
+
+            channel=FileChannel.open(Paths.get(file_name), options);
             buffer=new UnsafeBuffer(channel.map(FileChannel.MapMode.READ_WRITE, 0, buffer_length));
             rb=new ManyToOneRingBuffer(buffer);
         }
